\name{GFGM.copula-package}
\alias{GFGM.copula-package}
\docType{package}
\title{
\packageTitle{GFGM.copula}
}
\description{
\packageDescription{GFGM.copula}
}
%\details{

%}
\author{
\packageAuthor{GFGM.copula}

Maintainer: \packageMaintainer{GFGM.copula}
}
\references{
Shih and Emura (2016) Bivariate dependence measures and bivariate competing risks models under the generalized FGM copula, Statistical Papers, doi: 10.1007/s00362-016-0865-5.

Shih and Emura (2018) Likelihood inference for bivariate latent failure time models with competing risks udner the generalized FGM copula (in re-submission, Computational Statistics).
}

