\name{GFGM.copula-package}
\alias{GFGM.copula-package}
\docType{package}
\title{
\packageTitle{GFGM.copula}
}
\description{
Compute bivariate dependence measures and perform bivariate competing risks analysis under the generalized Farlie-Gumbel-Morgenstern (FGM) copula. See Shih and Emura (2016, 2018) for details.
}
\details{
The functions in this package are based on latent failure time models with competing risks in Shih and Emura (2018). However, they can be adapted to dependent censoring models in Emura and Chen (2018). See \code{MLE.GFGM.spline} for example.
}
\author{
\packageAuthor{GFGM.copula}

Maintainer: \packageMaintainer{GFGM.copula}
}
\references{
Shih J-H, Emura T (2016) Bivariate dependence measures and bivariate competing risks models under the generalized FGM copula, Statistical Papers, doi: 10.1007/s00362-016-0865-5.

Shih J-H, Emura T (2018) Likelihood-based inference for bivariate latent failure time models with competing risks udner the generalized FGM copula, Computational Statistics, doi: 10.1007/s00180-018-0804-0.

Emura T, Chen Y-H (2018) Analysis of Survival Data with Dependent Censoring, Copula-Based Approaches, JSS Research Series in Statistics, Springer, in press.
}

